[MOD] ProjectServiceImplTest 코드를 2개의 파일로 분류
